backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
python-dotenv==1.0.0
python-multipart==0.0.6
numpy==1.24.3
pandas==2.0.3
